requirements.txt
qrcode
django
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
Pillow